requirements.txt
streamlit>=1.32.0
pymongo>=4.6.0
pandas>=2.0.0
plotly>=5.22.0
python-dotenv>=1.0.0
